Go implementation: simple replace of math.rand with crypto.rand
Just an ugly replacement of rand.Intn() from math/rand with rand.Int()
from crypto/rand.
This could be done in a cleaner way. For example via an extra function
which returns the result fitting to a given range.

What still remains is the use of math/rand.Perm(). This still needs to
be addressed.